// services/api-service/src/auth/interfaces/auth.ts
import { User } from '../../users/entities';

export interface jwtPayload {
  userId: string;
  role: string;
}
export interface signUpReturn extends signInReturn {
  user: User;
}

export interface signInReturn {
  accessToken: string;
  refreshToken: string;
}

export interface SendVerificationData {
  to: string;
  subject: string;
  text: string;
}

export interface IVerification {
  email: string;
  token: string;
  type: string;
  expiredAt: Date;
}

export interface BasicReturnType<T> {
  success?: boolean;
  error?: boolean;
  data?: T;
  message?: string;
}

export interface ResendCodeDTO {
  email: string;
  type: string;
}

export interface GetQRCodeReturn {
  code: string;
}
export interface IUserFromHeaders {
  id: string;
  email: string;
  role: string;
  activatedAt: Date;
  refreshToken: string;
  settings: object;
  createdAt: Date;
  updatedAt: Date;
}